requirements.txt
numpy==1.25.2
opencv_python==4.7.0.72
opencv_python_headless==4.8.0.74
Pillow==10.0.0
ultralytics==8.0.109
